auth of deletion using policy

+++ larablog/app/Http/Controllers/PostController.php
@@ -9,7 +9,7 @@ class PostController extends Controller
 {
     public function index() {
 
-        $posts = Post::with(['user', 'likes'])->paginate(20); //all posts
+        $posts = Post::latest()->with(['user', 'likes'])->paginate(20); //all posts
         return view('posts.index', [
             'posts' => $posts
         ]);
@@ -26,6 +26,7 @@ class PostController extends Controller
     }
 
     public function destroy(Post $post) {
+        $this->authorize('delete', $post); //throws exception
         $post->delete();
 
         return back();
